+++ packages/wrangler/src/api/dev.ts
@@ -1,6 +1,6 @@
 import events from "node:events";
 import { fetch, Request } from "undici";
-import { startDev } from "../dev";
+import { startDev } from "../dev/start-dev";
 import { getDockerPath } from "../environment-variables/misc-variables";
 import { run } from "../experimental-flags";
 import { logger } from "../logger";
