+++ packages/wrangler/src/dev.ts
@@ -1,35 +1,18 @@
 import assert from "node:assert";
 import events from "node:events";
-import path from "node:path";
-import { bold, green } from "@cloudflare/cli/colors";
-import { generateContainerBuildId } from "@cloudflare/containers-shared";
 import { isWebContainer } from "@webcontainer/env";
-import dedent from "ts-dedent";
-import { DevEnv } from "./api";
-import { MultiworkerRuntimeController } from "./api/startDevWorker/MultiworkerRuntimeController";
-import { NoOpProxyController } from "./api/startDevWorker/NoOpProxyController";
-import { convertCfWorkerInitBindingsToBindings } from "./api/startDevWorker/utils";
 import { getAssetsOptions } from "./assets";
 import { configFileName, formatConfigSnippet } from "./config";
 import { createCommand } from "./core/create-command";
 import { validateRoutes } from "./deploy/deploy";
-import { validateNodeCompatMode } from "./deployment-bundle/node-compat";
 import { getVarsForDev } from "./dev/dev-vars";
-import registerDevHotKeys from "./dev/hotkeys";
-import { getRegistryPath } from "./environment-variables/misc-variables";
+import { startDev } from "./dev/start-dev";
 import { UserError } from "./errors";
 import { getFlag } from "./experimental-flags";
-import isInteractive from "./is-interactive";
 import { logger } from "./logger";
-import { getSiteAssetPaths } from "./sites";
-import { requireApiToken, requireAuth } from "./user";
-import {
-	collectKeyValues,
-	collectPlainTextVars,
-} from "./utils/collectKeyValues";
 import { mergeWithOverride } from "./utils/mergeWithOverride";
 import { getHostFromRoute } from "./zones";
-import type { AsyncHook, StartDevWorkerInput, Trigger } from "./api";
+import type { Trigger } from "./api";
 import type { Config, Environment } from "./config";
 import type {
 	EnvironmentNonInheritable,
@@ -46,9 +29,7 @@ import type {
 	CfService,
 	CfWorkerInit,
 } from "./deployment-bundle/worker";
-import type { CfAccount } from "./dev/create-worker-preview";
 import type { EnablePagesAssetsServiceBindingOptions } from "./miniflare-cli/types";
-import type { watch } from "chokidar";
 import type { Json } from "miniflare";
 
 export const dev = createCommand({
@@ -386,306 +367,6 @@ export type StartDevOptions = DevArguments &
 		containerEngine?: string;
 	};
 
-async function getPagesAssetsFetcher(
-	options: EnablePagesAssetsServiceBindingOptions | undefined
-): Promise<StartDevWorkerInput["bindings"] | undefined> {
-	if (options !== undefined) {
-		// `./miniflare-cli/assets` dynamically imports`@cloudflare/pages-shared/environment-polyfills`.
-		// `@cloudflare/pages-shared/environment-polyfills/types.ts` defines `global`
-		// augmentations that pollute the `import`-site's typing environment.
-		//
-		// We `require` instead of `import`ing here to avoid polluting the main
-		// `wrangler` TypeScript project with the `global` augmentations. This
-		// relies on the fact that `require` is untyped.
-		//
-		// eslint-disable-next-line @typescript-eslint/no-require-imports
-		const generateASSETSBinding = require("./miniflare-cli/assets").default;
-		return {
-			ASSETS: {
-				type: "fetcher",
-				fetcher: await generateASSETSBinding({
-					log: logger,
-					...options,
-				}),
-			},
-		};
-	}
-}
-
-async function setupDevEnv(
-	devEnv: DevEnv,
-	configPath: string | undefined,
-	auth: AsyncHook<CfAccount, [Pick<Config, "account_id">]>,
-	args: Partial<StartDevOptions> & { multiworkerPrimary?: boolean }
-) {
-	await devEnv.config.set(
-		{
-			name: args.name,
-			config: configPath,
-			entrypoint: args.script,
-			compatibilityDate: args.compatibilityDate,
-			compatibilityFlags: args.compatibilityFlags,
-			triggers: args.routes?.map<Extract<Trigger, { type: "route" }>>((r) => ({
-				type: "route",
-				pattern: r,
-			})),
-			env: args.env,
-			envFiles: args.envFile,
-			build: {
-				bundle: args.bundle !== undefined ? args.bundle : undefined,
-				define: collectKeyValues(args.define),
-				jsxFactory: args.jsxFactory,
-				jsxFragment: args.jsxFragment,
-				tsconfig: args.tsconfig,
-				minify: args.minify,
-				processEntrypoint: args.processEntrypoint,
-				additionalModules: args.additionalModules,
-				moduleRoot: args.moduleRoot,
-				moduleRules: args.rules,
-				nodejsCompatMode: (parsedConfig: Config) =>
-					validateNodeCompatMode(
-						args.compatibilityDate ?? parsedConfig.compatibility_date,
-						args.compatibilityFlags ?? parsedConfig.compatibility_flags ?? [],
-						{
-							noBundle: args.noBundle ?? parsedConfig.no_bundle,
-						}
-					),
-			},
-			bindings: {
-				...(await getPagesAssetsFetcher(args.enablePagesAssetsServiceBinding)),
-				...collectPlainTextVars(args.var),
-				...convertCfWorkerInitBindingsToBindings({
-					kv_namespaces: args.kv,
-					vars: args.vars,
-					send_email: undefined,
-					wasm_modules: undefined,
-					text_blobs: undefined,
-					browser: undefined,
-					ai: args.ai,
-					images: undefined,
-					version_metadata: args.version_metadata,
-					data_blobs: undefined,
-					durable_objects: { bindings: args.durableObjects ?? [] },
-					workflows: undefined,
-					queues: undefined,
-					r2_buckets: args.r2,
-					d1_databases: args.d1Databases,
-					vectorize: undefined,
-					hyperdrive: undefined,
-					secrets_store_secrets: undefined,
-					unsafe_hello_world: undefined,
-					services: args.services,
-					analytics_engine_datasets: undefined,
-					dispatch_namespaces: undefined,
-					mtls_certificates: undefined,
-					pipelines: undefined,
-					logfwdr: undefined,
-					unsafe: undefined,
-					assets: undefined,
-				}),
-			},
-			dev: {
-				auth,
-				remote: !args.forceLocal && args.remote,
-				server: {
-					hostname: args.ip,
-					port: args.port,
-					secure:
-						args.localProtocol === undefined
-							? undefined
-							: args.localProtocol === "https",
-					httpsCertPath: args.httpsCertPath,
-					httpsKeyPath: args.httpsKeyPath,
-				},
-				inspector: {
-					port: args.inspectorPort,
-				},
-				origin: {
-					hostname: args.host ?? args.localUpstream,
-					secure:
-						args.upstreamProtocol === undefined
-							? undefined
-							: args.upstreamProtocol === "https",
-				},
-				persist: args.persistTo,
-				liveReload: args.liveReload,
-				testScheduled: args.testScheduled,
-				logLevel: args.logLevel,
-				registry: args.disableDevRegistry ? undefined : getRegistryPath(),
-				multiworkerPrimary: args.multiworkerPrimary,
-				enableContainers: args.enableContainers,
-				dockerPath: args.dockerPath,
-				// initialise with a random id
-				containerBuildId: generateContainerBuildId(),
-			},
-			legacy: {
-				site: (configParam) => {
-					const legacyAssetPaths = getResolvedSiteAssetPaths(args, configParam);
-					return Boolean(args.site || configParam.site) && legacyAssetPaths
-						? {
-								bucket: path.join(
-									legacyAssetPaths.baseDirectory,
-									legacyAssetPaths?.assetDirectory
-								),
-								include: legacyAssetPaths.includePatterns,
-								exclude: legacyAssetPaths.excludePatterns,
-							}
-						: undefined;
-				},
-				enableServiceEnvironments: !(args.legacyEnv ?? true),
-			},
-			assets: args.assets,
-		} satisfies StartDevWorkerInput,
-		true
-	);
-	return devEnv;
-}
-
-export async function startDev(args: StartDevOptions) {
-	let configFileWatcher: ReturnType<typeof watch> | undefined;
-	let assetsWatcher: ReturnType<typeof watch> | undefined;
-	let devEnv: DevEnv | DevEnv[] | undefined;
-	let teardownRegistryPromise:
-		| Promise<(name?: string) => Promise<void>>
-		| undefined;
-
-	let unregisterHotKeys: (() => void) | undefined;
-	try {
-		if (args.logLevel) {
-			logger.loggerLevel = args.logLevel;
-		}
-
-		const authHook: AsyncHook<CfAccount, [Pick<Config, "account_id">]> = async (
-			config
-		) => {
-			const hotkeysDisplayed = !!unregisterHotKeys;
-			let accountId = args.accountId;
-			if (!accountId) {
-				unregisterHotKeys?.();
-				accountId = await requireAuth(config);
-				if (hotkeysDisplayed) {
-					assert(devEnv !== undefined);
-					unregisterHotKeys = registerDevHotKeys(
-						Array.isArray(devEnv) ? devEnv : [devEnv],
-						args,
-						false
-					);
-				}
-			}
-			return {
-				accountId,
-				apiToken: requireApiToken(),
-			};
-		};
-
-		if (args.remote) {
-			logger.log(
-				bold(
-					dedent`
-						Support for remote bindings in ${green("`wrangler dev`")} is now available as a replacement for ${green("`wrangler dev --remote`")}. Try it out now by running ${green("`wrangler dev`")} with the ${green("`remote`")} option enabled on your resources and let us know how it goes!
-						This gives you access to remote resources in development while retaining all the usual benefits of local dev: fast iteration speed, breakpoint debugging, and more.
-
-						Refer to https://developers.cloudflare.com/workers/development-testing/#remote-bindings for more information.`
-				)
-			);
-		}
-
-		if (Array.isArray(args.config)) {
-			const runtime = new MultiworkerRuntimeController(args.config.length);
-
-			const primaryDevEnv = new DevEnv({ runtimes: [runtime] });
-
-			// Set up the primary DevEnv (the one that the ProxyController will connect to)
-			devEnv = [
-				await setupDevEnv(primaryDevEnv, args.config[0], authHook, {
-					...args,
-					multiworkerPrimary: true,
-				}),
-			];
-
-			// Set up all auxiliary DevEnvs
-			devEnv.push(
-				...(await Promise.all(
-					(args.config as string[]).slice(1).map((c) => {
-						return setupDevEnv(
-							new DevEnv({
-								runtimes: [runtime],
-								proxy: new NoOpProxyController(),
-							}),
-							c,
-							authHook,
-							{
-								disableDevRegistry: args.disableDevRegistry,
-								multiworkerPrimary: false,
-							}
-						);
-					})
-				))
-			);
-			if (isInteractive() && args.showInteractiveDevSession !== false) {
-				unregisterHotKeys = registerDevHotKeys(devEnv, args);
-			}
-		} else {
-			devEnv = new DevEnv();
-
-			await setupDevEnv(devEnv, args.config, authHook, args);
-
-			if (isInteractive() && args.showInteractiveDevSession !== false) {
-				unregisterHotKeys = registerDevHotKeys([devEnv], args);
-			}
-		}
-
-		const [primaryDevEnv, ...secondary] = Array.isArray(devEnv)
-			? devEnv
-			: [devEnv];
-
-		// The ProxyWorker will have a stable host and port, so only listen for the first update
-		void primaryDevEnv.proxy.ready.promise.then(({ url }) => {
-			if (args.onReady) {
-				args.onReady(url.hostname, parseInt(url.port));
-			}
-
-			if (
-				(args.enableIpc || !args.onReady) &&
-				process.send &&
-				typeof vitest === "undefined"
-			) {
-				process.send(
-					JSON.stringify({
-						event: "DEV_SERVER_READY",
-						ip: url.hostname,
-						port: parseInt(url.port),
-					})
-				);
-			}
-		});
-
-		return {
-			devEnv: primaryDevEnv,
-			secondary,
-			unregisterHotKeys,
-			teardownRegistryPromise,
-		};
-	} catch (e) {
-		await Promise.allSettled([
-			configFileWatcher?.close(),
-			assetsWatcher?.close(),
-			...(Array.isArray(devEnv)
-				? devEnv.map((d) => d.teardown())
-				: [devEnv?.teardown()]),
-			(async () => {
-				if (teardownRegistryPromise) {
-					assert(devEnv === undefined || !Array.isArray(devEnv));
-					const teardownRegistry = await teardownRegistryPromise;
-					await teardownRegistry(devEnv?.config.latestConfig?.name);
-				}
-				unregisterHotKeys?.();
-			})(),
-		]);
-		throw e;
-	}
-}
-
 /**
  * mask anything that was overridden in .dev.vars
  * so that we don't log potential secrets into the terminal
@@ -770,18 +451,6 @@ export function getInferredHost(
 		}
 		return host;
 	}
-}
-
-function getResolvedSiteAssetPaths(
-	args: Partial<StartDevOptions>,
-	configParam: Config
-) {
-	return getSiteAssetPaths(
-		configParam,
-		args.site,
-		args.siteInclude,
-		args.siteExclude
-	);
 }
 
 /**
